Can Mine Icon
A green check and red x were added to the end of the pick power line to indicate whether the player's current held pickaxe's power is strong enough to mine the tile.

The recommended pickaxe only shows/cycles when the current pickaxe cannot mine the tile. In other words only show recommendations if they need them